+++ lib/keycloak-admin/client/client.rb
@@ -2,6 +2,7 @@ require "uri"
 
 module KeycloakAdmin
   class Client
+    include PathEncoding
 
     def initialize(configuration)
       @configuration = configuration